requirements.txt

# Core Framework
fastapi==0.115.8
uvicorn==0.34.0
jinja2==3.1.4

# Database
sqlalchemy==2.0.37
asyncpg==0.30.0
alembic==1.14.1

# FastAPI Extensions
python-multipart==0.0.20
aiofiles==24.1.0

# Configuration & Validation
python-dotenv==1.0.1
pydantic==2.9.1
pydantic-settings==2.7.1
aiofiles==24.1.0

# Unit Testing
pytest-mock
pytest-asyncioreportlab==4.0.8
